feat(core): print normalized generate command to the terminal
