+++ packages/nx/src/tasks-runner/utils.spec.ts
@@ -1,4 +1,4 @@
-import { getOutputsForTargetAndConfiguration, unparse } from './utils';
+import { getOutputsForTargetAndConfiguration } from './utils';
 import { ProjectGraphProjectNode } from '../config/project-graph';
 
 describe('utils', () => {
@@ -199,75 +199,6 @@ describe('utils', () => {
           'root-myapp/public',
         ]);
       });
-    });
-  });
-
-  describe('unparse', () => {
-    it('should unparse options whose values are primitives', () => {
-      const options = {
-        boolean1: false,
-        boolean2: true,
-        number: 4,
-        string: 'foo',
-        'empty-string': '',
-        ignore: null,
-      };
-
-      expect(unparse(options)).toEqual([
-        '--no-boolean1',
-        '--boolean2',
-        '--number=4',
-        '--string=foo',
-        '--empty-string=',
-      ]);
-    });
-
-    it('should unparse options whose values are arrays', () => {
-      const options = {
-        array1: [1, 2],
-        array2: [3, 4],
-      };
-
-      expect(unparse(options)).toEqual([
-        '--array1=1',
-        '--array1=2',
-        '--array2=3',
-        '--array2=4',
-      ]);
-    });
-
-    it('should unparse options whose values are objects', () => {
-      const options = {
-        foo: {
-          x: 'x',
-          y: 'y',
-          w: [1, 2],
-          z: [3, 4],
-        },
-      };
-
-      expect(unparse(options)).toEqual([
-        '--foo.x=x',
-        '--foo.y=y',
-        '--foo.w=1',
-        '--foo.w=2',
-        '--foo.z=3',
-        '--foo.z=4',
-      ]);
-    });
-
-    it('should quote string values with space(s)', () => {
-      const options = {
-        string1: 'one',
-        string2: 'one two',
-        string3: 'one two three',
-      };
-
-      expect(unparse(options)).toEqual([
-        '--string1=one',
-        '--string2="one two"',
-        '--string3="one two three"',
-      ]);
     });
   });
 });
