+++ packages/nx/src/tasks-runner/utils.ts
@@ -1,4 +1,3 @@
-import { flatten } from 'flat';
 import { output } from '../utils/output';
 import { Workspaces } from '../config/workspaces';
 import { mergeNpmScriptsWithTargets } from '../utils/project-graph-utils';
@@ -13,6 +12,7 @@ import { getPackageManagerCommand } from '../utils/package-manager';
 import { ProjectGraph, ProjectGraphProjectNode } from '../config/project-graph';
 import { TargetDependencyConfig } from '../config/workspace-json-project-json';
 import { workspaceRoot } from '../utils/app-root';
+import { unparse } from '../utils/params';
 
 export function getCommandAsString(task: Task) {
   const execCommand = getPackageManagerCommand().exec;
@@ -107,47 +107,6 @@ export function getOutputsForTargetAndConfiguration(
   } else {
     return [];
   }
-}
-
-export function unparse(options: Object): string[] {
-  const unparsed = [];
-  for (const key of Object.keys(options)) {
-    const value = options[key];
-    unparseOption(key, value, unparsed);
-  }
-
-  return unparsed;
-}
-
-function unparseOption(key: string, value: any, unparsed: string[]) {
-  if (value === true) {
-    unparsed.push(`--${key}`);
-  } else if (value === false) {
-    unparsed.push(`--no-${key}`);
-  } else if (Array.isArray(value)) {
-    value.forEach((item) => unparseOption(key, item, unparsed));
-  } else if (Object.prototype.toString.call(value) === '[object Object]') {
-    const flattened = flatten<any, any>(value, { safe: true });
-    for (const flattenedKey in flattened) {
-      unparseOption(
-        `${key}.${flattenedKey}`,
-        flattened[flattenedKey],
-        unparsed
-      );
-    }
-  } else if (
-    typeof value === 'string' &&
-    stringShouldBeWrappedIntoQuotes(value)
-  ) {
-    const sanitized = value.replace(/"/g, String.raw`\"`);
-    unparsed.push(`--${key}="${sanitized}"`);
-  } else if (value != null) {
-    unparsed.push(`--${key}=${value}`);
-  }
-}
-
-function stringShouldBeWrappedIntoQuotes(str: string) {
-  return str.includes(' ') || str.includes('{') || str.includes('"');
 }
 
 function interpolateOutputs(template: string, data: any): string {
